# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Clicky-Store API
  version: 0.1.0
  description: REST API for the Clicky-Store educational e-commerce project.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /auth/register:
    post:
      summary: Register a customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered user and bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /auth/login:
    post:
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged-in user and bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /me:
    get:
      summary: Get the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
  /products:
    get:
      summary: List public products
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
...
  /cart:
    get:
      summary: Get the authenticated user's cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  cart:
                    $ref: "#/components/schemas/Cart"
        "401":
          $ref: "#/components/responses/Error"
  /cart/items:
    post:
      summary: Add an item to the authenticated user's cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemRequest"
      responses:
        "201":
          description: Updated cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  cart:
                    $ref: "#/components/schemas/Cart"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /orders:
    get:
      summary: List the authenticated user's orders
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/Orders"
    post:
      summary: Create an order from the authenticated user's cart
      description: Creates a pending simulated payment, decreases stock, snapshots order items, and clears the cart transactionally.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paymentMethod:
                  type: string
                  example: simulation
      responses:
        "201":
          $ref: "#/components/responses/Order"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /admin/products:
    get:
      summary: List products as admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Product list
    post:
      summary: Create a product as admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "201":
          description: Created product
        "403":
          $ref: "#/components/responses/Error"
...
  /admin/orders:
    get:
      summary: List all orders as admin
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/Orders"
  /admin/users:
    get:
      summary: List users as admin
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum:
              - admin
              - customer
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: User list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: JSON error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Order:
      description: Order
      content:
        application/json:
          schema:
            type: object
            properties:
              order:
                $ref: "#/components/schemas/Order"
    Orders:
      description: Order list
      content:
        application/json:
          schema:
            type: object
            properties:
              orders:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - admin
            - customer
        createdAt:
          type: string
          format: date-time
    ProductRequest:
      type: object
      required:
        - name
        - slug
        - description
        - category
        - priceCents
        - currency
        - dpi
        - stock
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
        priceCents:
          type: integer
          minimum: 1
        currency:
          type: string
          example: PLN
        dpi:
          type: integer
          minimum: 1
        wireless:
          type: boolean
        ergonomic:
          type: boolean
        stock:
          type: integer
          minimum: 0
        imageUrl:
          type: string
    Product:
      allOf:
        - $ref: "#/components/schemas/ProductRequest"
        - type: object
          properties:
            id:
              type: string
            images:
              type: array
              items:
                $ref: "#/components/schemas/ProductImage"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    ProductImage:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        url:
          type: string
          example: /assets/products/viper-x1.svg
        altText:
          type: string
        sortOrder:
          type: integer
          minimum: 0
        isPrimary:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    CartItemRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1
    Cart:
      type: object
      properties:
        userId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartLine"
        totalCents:
          type: integer
        currency:
          type: string
    CartLine:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        subtotalCents:
          type: integer
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalCents:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum:
            - pending
            - confirmed
            - payment_failed
        paymentStatus:
          type: string
          enum:
            - pending
            - paid
            - failed
        paymentMethod:
          type: string
        createdAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        quantity:
          type: integer
        unitPriceCents:
          type: integer
        subtotalCents:
          type: integer